cameraFragment added, need performance work
